requirements.txt
uvicorn
fastapi
pydantic
pydantic[email]
email-validator
supabase
python-dotenv
requests
PyPDF2
python-docx
websockets
streamlit